Update GithubAssignment2.java
I wrote a static method that takes an Array of integers as an input. After that I used a loop to find the sum of the terms of the array and divided them to the term number to find the average. Then I used another loop to find the differences between the terms and the average and stored them in an array. Lastly I displayed the array of differences